remove old prototype
