+++ ui/src/app.html
@@ -108,9 +108,8 @@
             document.querySelector('.patchies-docs') ||
             // debug pages (/debug/)
             document.querySelector('.debug-page') ||
-            // sparks pages (/sparks/)
-            document.querySelector('.sparks-page') ||
-            document.querySelector('.mood-page');
+            // sparks page (/sparks/)
+            document.querySelector('.sparks-page');
 
           if (flowContainer) {
             // Patcher UI is mounted, hide loader after brief delay
